@@ -1,3 +1,3 @@
 module golang.org/dl
 
-go 1.11
+go 1.18
